# src/app/api/v1/games/[gameId]/route.ts
import { NextRequest, NextResponse } from 'next/server';
import { gameService } from '@/services/api.service';

// GET /api/v1/games/{gameId} - Get game by ID
export async function GET(
  request: NextRequest,
  { params }: { params: Promise<{ gameId: string }> }
) {
  try {
    const { gameId } = await params;
    
    const game = await gameService.getGame(gameId);
    
    if (!game) {
      return NextResponse.json(
        { status: 404, message: `Game with ID '${gameId}' not found` },
        { status: 404 }
      );
    }
    
    return NextResponse.json(game, { status: 200 });
  } catch (error) {
    console.error('Error retrieving game:', error);
    return NextResponse.json(
      { status: 500, message: 'Internal server error occurred while processing the request' },
      { status: 500 }
    );
  }
}

// POST /api/v1/games/{gameId} - Start a game by ID
export async function POST(
  request: NextRequest,
  { params }: { params: Promise<{ gameId: string }> }
) {
  try {
    const { gameId } = await params;
    
    const game = await gameService.getGame(gameId);
    
    if (!game) {
      return NextResponse.json(
        { status: 404, message: `Game with ID '${gameId}' not found` },
        { status: 404 }
      );
    }
    
    // Check if the game has already been started
    if (game.scenes && game.scenes.length > 0) {
      return NextResponse.json(
        { status: 409, message: 'Game has already been started' },
        { status: 409 }
      );
    }
    
    const success = await gameService.startGame(gameId);
    
    if (!success) {
      return NextResponse.json(
        { status: 500, message: 'Failed to start the game' },
        { status: 500 }
      );
    }
    
    return NextResponse.json({ message: 'Game started successfully' }, { status: 200 });
  } catch (error) {
    console.error('Error starting game:', error);
    return NextResponse.json(
      { status: 500, message: 'Internal server error occurred while processing the request' },
      { status: 500 }
    );
  }
} 